Move get manifest to context package
